Make scripts without extension throw correct exceptions
